+++ src/main/java/com/osudpotro/posmaster/purchase/purchasecart/PurchaseCartItemDto.java
@@ -4,12 +4,14 @@ import com.osudpotro.posmaster.product.ProductDetailDto;
 import com.osudpotro.posmaster.product.ProductDto;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PurchaseCartItemDto {
     private Long id;
     private ProductDto product;
     private ProductDetailDto productDetail;
-    private Double purchasePrice;
-    private Long purchaseQty;
-    private Double purchaseLinePrice;
+    private BigDecimal purchasePrice;
+    private Integer purchaseQty;
+    private BigDecimal purchaseLinePrice;
 }
